moved BookContext
