+++ BooksStore/API/Startup.cs
@@ -6,8 +6,8 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.OpenApi.Models;
+using Model;
 using Model.Handlers;
-using Model.Models;
 using Service;
 using Service.BookRepos;
 using Service.PainterRepos;
